Add authorization for event deletion

+++ backend/Services/eventsService.js
@@ -297,14 +297,19 @@ class EventsService {
         return event;
     }
 
-    async deleteEvent(eventId) {
+    async deleteEvent(eventId, organizerID) {
         const event = await eventsModel.findById(eventId);
         if (!event) {
             throw new Error("Event not found");
         }
 
-        // Delete the event
-        await eventsModel.deleteOne({ _id: eventId });
+        // Only the organizer can delete the event
+        if (event.eventOrganizer.toString() !== organizerID) {
+            throw new Error("You are not authorized to delete this event");
+        }else{
+            // Delete the event
+            await eventsModel.deleteOne({ _id: eventId });
+        }
 
         // Delete the event from the user's participatedEvents array
         for (const participant of event.eventParticipant) {
